Added TextBox option
Added textbox and a button to summon a textbox beside the other Nodes. The textbox moves vertically when width is overloaded and add scrolling effect when text overflow the div. Met issues when trying to implement automatic expantion of the textbox to fit the input text so this function is not added yet.